Removed mocked gRPC manager. Mock data moved to separate file
